requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
motor==3.3.2  # Motor requires pymongo>=4.5
pymongo==4.5  # Ensure compatibility
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
PyJWT==2.8.0
pydantic==2.5.3
email-validator==2.1.0.post1
